+++ src/Archive/ArchiveUnpacker.cpp
@@ -1,6 +1,6 @@
 #include "ArchiveUnpacker.h"
 #include "../XFile.h"
-#include "../Stream/SeekableReader.h"
+#include "../Stream/BiDirectionalSeekableReader.h"
 #include "../Stream/FileWriter.h"
 #include <array>
 #include <string>
