+++ src/Stream/BiDirectionalSeekableReader.h
