+++ src/Stream/BiDirectionalSeekableWriter.h
